Added other endpoints and also work error

+++ src/main/java/org/polymath/noteapi/controller/NotesController.java
@@ -1,9 +1,11 @@
 package org.polymath.noteapi.controller;
 
+import io.jsonwebtoken.MalformedJwtException;
 import org.polymath.noteapi.models.Notes;
 import org.polymath.noteapi.models.Users;
 import org.polymath.noteapi.repositories.NotesRepo;
 import org.polymath.noteapi.repositories.UserRepo;
+import org.polymath.noteapi.service.JWTService;
 import org.polymath.noteapi.util.ApiError;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -12,10 +14,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
 
 @RestController
 @RequestMapping("/api/notes")
@@ -26,32 +25,53 @@ public class NotesController {
 
     @Autowired
     private UserRepo userRepo;
+    @Autowired
+    private JWTService jwtService;
 
-    @GetMapping("")
-    public List<Notes> getNotes() {
-        return notesRepo.findAll();
-    }
 //    @GetMapping("")
-//    public ResponseEntity<?> getAllUserNotes(@PathVariable UUID userId) {
-//        try {
-//            Optional<Notes> notes = notesRepo.findAllByUserId(userId);
-//            if (notes.isPresent()) {
-//                return new ResponseEntity<>(notes.get(), HttpStatus.OK);
-//            }else {
-//                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-//            }
-//        } catch (Exception e) {
-//            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ApiError(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage(), null));
-//        }
-//
+//    public List<Notes> getNotes() {
+//        return notesRepo.findAll();
 //    }
+    @GetMapping("")
+    public ResponseEntity<?> getAllUserNotes(@RequestHeader("Authorization") String authHeader) {
+        try {
+           // System.out.println(authHeader);
+            String token = authHeader.substring(7);
+            UUID userId = UUID.fromString(jwtService.extractUserId(token));
+            List<Notes> notes = notesRepo.findAllByUserId(userId);
+            if (notes.isEmpty()) {
+                return ResponseEntity.ok(Collections.emptyList());
+            }
+            return ResponseEntity.ok(notes);
+
+
+        }catch (MalformedJwtException e) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ApiError(HttpStatus.UNAUTHORIZED,e.getMessage(),null));
+        }catch (IllegalArgumentException e) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(new ApiError(HttpStatus.BAD_REQUEST, "Invalid Authorization header format", null));
+        }
+        catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ApiError(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage(), null));
+        }
+
+    }
     @PostMapping("/save")
-    public Notes saveNote(@RequestBody Notes note) {
-        UUID userId = note.getUserId();
-        Users user = userRepo.findById(userId).orElseThrow(() -> new RuntimeException("User Not Found"));
-        note.setUserId(userId);
-        note.setAuthor(user.getFirstName()+" "+user.getLastName());
-        return notesRepo.save(note);
+    public ResponseEntity<?> saveNote(@RequestBody Notes note) {
+        try {
+            UUID userId = note.getUserId();
+            Users user = userRepo.findById(userId).orElseThrow(() -> new RuntimeException("User Not Found"));
+            note.setUserId(userId);
+            note.setLastModified(LocalDateTime.now());
+            note.setAuthor(user.getFirstName() + " " + user.getLastName());
+             notesRepo.save(note);
+             return ResponseEntity.ok(note);
+        }catch (MalformedJwtException e) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ApiError(HttpStatus.UNAUTHORIZED,e.getMessage(),null));
+        }
+        catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ApiError(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage(), null));
+        }
     }
 
 
@@ -61,18 +81,32 @@ public class NotesController {
     }
 
     @GetMapping("/search")
-    public ResponseEntity<?> searchAllByTitle(@RequestParam String title) {
+    public ResponseEntity<?> searchAllByTitleOrContentOrTag(@RequestParam(required = false) String title, @RequestParam(required = false) String content, @RequestParam(required = false) List<String> tags,@RequestHeader("Authorization") String authHeader) {
+
        try {
-           if(title==null || title.isEmpty()) {
-               return ResponseEntity.badRequest().body(new ApiError(HttpStatus.BAD_REQUEST,"title must be null or empty",null));
+           String token = authHeader.substring(7);
+           UUID userId = UUID.fromString(jwtService.extractUserId(token));
+
+           if ((title == null || title.isEmpty()) &&
+                   (content == null || content.isEmpty()) &&
+                   (tags == null || tags.isEmpty())) {
+               return ResponseEntity.badRequest().body(new ApiError(HttpStatus.BAD_REQUEST, "At least one search parameter must be provided", null));
            }
-           Optional<Notes> notes = notesRepo.findAllByTitleIsContainingIgnoreCase(title);
-           if(notes.isPresent()) {
-               return ResponseEntity.ok(notes.get());
-           }else {
-               return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ApiError(HttpStatus.NOT_FOUND,"No note(s) found for this title",null));
+
+           if (tags == null) {
+               tags = Collections.emptyList();
            }
-       }catch (Exception e) {
+//           List<Notes> notesList = notesRepo.findAllByUserId(userId);
+           System.out.println(userId);
+           List<Notes> notes = notesRepo.findNotesByTitleOrContentOrTags(title,content,tags,userId);
+          if (notes.isEmpty()) {
+              return ResponseEntity.ok(Collections.emptyList());
+          }
+          return ResponseEntity.ok(notes);
+       }catch (MalformedJwtException e) {
+           return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ApiError(HttpStatus.UNAUTHORIZED,e.getMessage(),null));
+       }
+       catch (Exception e) {
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ApiError(HttpStatus.INTERNAL_SERVER_ERROR,e.getMessage(),null));
        }
     }
@@ -89,8 +123,12 @@ public class NotesController {
             existingNote.setTitle(note.getTitle());
             existingNote.setContent(note.getContent());
             existingNote.setTags(note.getTags());
+            existingNote.setLastModified(LocalDateTime.now());
+//            existingNote.setAuthor(note.;
             notesRepo.save(existingNote);
             return ResponseEntity.ok(existingNote);
+        }catch (MalformedJwtException e){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ApiError(HttpStatus.UNAUTHORIZED,e.getMessage(),null));
         } catch (Exception e) {
            return ResponseEntity.badRequest().body(new ApiError(HttpStatus.INTERNAL_SERVER_ERROR,e.getMessage(),null));
         }
@@ -98,9 +136,16 @@ public class NotesController {
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<?> deleteNote(@PathVariable UUID id) {
         try {
+            Notes notes = notesRepo.findNoteById(id);
+            if(notes == null) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ApiError(HttpStatus.NOT_FOUND,"Note not found",null));
+            }
            notesRepo.deleteById(id);
-           return ResponseEntity.ok().build();
-        } catch (Exception e) {
+           return ResponseEntity.status(HttpStatus.OK).body(notes);
+        }catch (MalformedJwtException e){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ApiError(HttpStatus.UNAUTHORIZED,e.getMessage(),null));
+        }
+        catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ApiError(HttpStatus.INTERNAL_SERVER_ERROR,e.getMessage(),null));
         }
     }
